Use new generateClassName function for each WuiThemeProvider
During server side rendering, the same generateClassName function
was being used to generate class names across different requests.
This worked fine during static generation, but for server side
rendering the counter in the class name would increase arbitrarily
and not match the client side render. This ensures that a new
generateClassName is passed for each request and memoizes it so
that it does not change on rerender.